Address compiler warnings

+++ include/Color.h
@@ -6,6 +6,8 @@
 #define SIMPLEECS_API __declspec(dllimport)
 #endif
 
+#include <cstdint>
+
 namespace SimpleECS
 {
 	/**
@@ -15,10 +17,10 @@ namespace SimpleECS
 
 	public:
 		// TODO: Hexadecimal constructor
-		Color(int r, int g, int b, int a) : r(r), g(g), b(b), a(a) {};
-		Color(int r, int g, int b) : r(r), g(g), b(b), a(1) {}
+		Color(uint8_t r, uint8_t g, uint8_t b, uint8_t a) : r(r), g(g), b(b), a(a) {};
+		Color(uint8_t r, uint8_t g, uint8_t b) : r(r), g(g), b(b), a(1) {}
 		Color() : r(0), g(0), b(0), a(1) {};
 
-		int r, g, b, a;
+		uint8_t r, g, b, a;
 	};
 }
